Further cleanup/re-org
